refactor: remove unnecessary dependencies and improve task configurations in turbo.json
- Updated build and dev tasks to depend only on the necessary build tasks, removing redundant db:generate dependencies.
- Enhanced db:generate task configuration by specifying outputs for better clarity and management.
- Ensured that the start task maintains its dependency on the build task while keeping cache disabled.
