+++ packages/database/package.json
@@ -2,27 +2,24 @@
   "name": "@repo/database",
   "version": "1.0.0",
   "type": "module",
-  "main": "dist/src/client.js",
-  "types": "dist/src/client.d.ts",
+  "main": "dist/src/index.js",
+  "types": "dist/src/index.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/src/client.js",
-      "default": "./dist/src/client.js",
-      "types": "./dist/src/client.d.ts"
+      "import": "./dist/src/index.js",
+      "default": "./dist/src/index.js",
+      "types": "./dist/src/index.d.ts"
     }
   },
   "scripts": {
     "db:push": "prisma db push",
     "db:seed": "tsx src/seed/index.ts",
     "format": "prisma format",
-    "db:generate": "prisma generate",
-    "postgenerate": "node ./scripts/fix-prisma-imports-robust.cjs",
+    "db:generate": "prisma generate && node ./scripts/fix-prisma-imports-robust.cjs",
     "lint": "eslint . --max-warnings 0",
-    "prebuild": "npm run db:generate",
-    "predev": "npm run db:generate",
-    "build": "tsc -p tsconfig.build.json",
+    "build": "pnpm db:generate && tsc -p tsconfig.build.json",
     "build:watch": "tsc -p tsconfig.build.json -w",
-    "postbuild": "node ./scripts/fix-prisma-imports-robust.cjs",
+    "dev": "pnpm db:generate",
     "studio": "prisma studio"
   },
   "dependencies": {
