+++ packages/database/src/index.ts
@@ -1,0 +1,19 @@
+// ===== Prisma Client & Types =====
+// Re-export everything from generated Prisma client
+export { PrismaClient, $Enums, Prisma } from "../generated/prisma/client.js";
+
+// Export all types (models and namespace types)
+export type * from "../generated/prisma/client.js";
+
+// Explicit re-exports for commonly used types
+export type {
+  Category,
+  Product,
+  User,
+  Config,
+  C,
+} from "../generated/prisma/client.js";
+
+// ===== Extended Client =====
+// Export the extended client instance and type
+export { prisma, type ExtendedPrismaClient } from "./client.js";
