+++ packages/database/src/seed/products.seed.ts
@@ -1,4 +1,4 @@
-import { prisma, Prisma } from "../client.js";
+import { prisma, Prisma } from "../index.js";
 import { CategoryName, ReadOutput } from "./categories.seed.js";
 
 export type ProductCreateWithoutCategoryInput =
