+++ packages/database/tsconfig.json
@@ -9,12 +9,6 @@
     "declarationMap": true,
     "outDir": "dist"
   },
-  "include": ["src/**/*.ts", "prisma.config.ts"],
-  "exclude": [
-    "node_modules",
-    "dist",
-    "**/*.test.ts",
-    ".turbo/**",
-    "generated/**"
-  ]
+  "include": ["src/**/*.ts", "prisma.config.ts", "generated/**/*.ts"],
+  "exclude": ["node_modules", "dist", "**/*.test.ts", ".turbo/**"]
 }
