+++ packages/domain/tsconfig.json
@@ -6,17 +6,16 @@
     "target": "ES2022",
     "lib": ["ES2022"],
     "rootDir": "src",
-    "outDir": "dist",
+    "outDir": "dist/src",
     "declaration": true,
     "declarationMap": true,
     "skipLibCheck": true,
     "strict": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
-    "composite": true,
-    "disableSourceOfProjectReferenceRedirect": true
+    "composite": true
   },
   "include": ["src/**/*.ts"],
   "exclude": ["dist", "node_modules"],
-  "references": [{ "path": "../database" }]
+  "references": [{ "path": "../database/tsconfig.build.json" }]
 }
