+++ turbo.json
@@ -2,24 +2,23 @@
   "$schema": "https://turborepo.com/schema.json",
   "tasks": {
     "build": {
-      "dependsOn": ["^build", "^db:generate"],
+      "dependsOn": ["^build"],
       "inputs": ["$TURBO_DEFAULT$", ".env*"],
       "outputs": ["dist/**"],
       "env": ["DATABASE_URL"]
     },
     "dev": {
-      "dependsOn": ["^db:generate"],
+      "dependsOn": ["^build"],
       "cache": false,
       "persistent": true
     },
     "start": {
       "dependsOn": ["^build"],
-
       "cache": false
     },
     "db:generate": {
-      "dependsOn": ["^db:generate"],
-      "cache": false
+      "cache": false,
+      "outputs": ["generated/**"]
     },
     "db:push": {
       "cache": false
